+++ assets/eip-7265/src/utils/LimiterLib.sol
@@ -60,10 +60,10 @@ library LimiterLib {
             return;
         }
 
-        uint256 currentTickTimestamp = block.timestamp - (block.timestamp % tickLength);
+        uint32 currentTickTimestamp = uint32(block.timestamp - (block.timestamp % tickLength));
         limiter.liqInPeriod += amount;
 
-        uint256 listHead = limiter.listHead;
+        uint32 listHead = limiter.listHead;
         if (listHead == 0) {
             // if there is no head, set the head to the new inflow
             limiter.listHead = currentTickTimestamp;
@@ -81,7 +81,7 @@ library LimiterLib {
             }
 
             // check if tail is the same as block.timestamp (multiple txs in same block)
-            uint256 listTail = limiter.listTail;
+            uint32 listTail = limiter.listTail;
             if (listTail == currentTickTimestamp) {
                 // add amount
                 limiter.listNodes[currentTickTimestamp].amount += amount;
@@ -108,7 +108,7 @@ library LimiterLib {
         uint256 withdrawalPeriod,
         uint256 totalIters
     ) internal {
-        uint256 currentHead = limiter.listHead;
+        uint32 currentHead = limiter.listHead;
         int256 totalChange = 0;
         uint256 iter = 0;
 
@@ -129,8 +129,8 @@ library LimiterLib {
 
         if (currentHead == 0) {
             // If the list is empty, set the tail and head to current times
-            limiter.listHead = block.timestamp;
-            limiter.listTail = block.timestamp;
+            limiter.listHead = uint32(block.timestamp);
+            limiter.listTail = uint32(block.timestamp);
         } else {
             limiter.listHead = currentHead;
         }
